1 (#2)
* fix: mouse right-click anomaly and warning caused by permission monitoring

* chore: download client to cache dir before copying after full install
This approach is intended to ensure that if the user forcefully closes the application during the game installation or if an unexpected bug occurs during the installation process, an incomplete version won't be mistakenly recognized by the VersionsManager as a valid one. At the same time, this solution effectively replaces the clumsy method of using version ignore flags to make the VersionsManager skip that version.

* fix: improved handling of undownloadable files

* fix: player model issue with offline account skins

* chore: some warnings about offline account skins

* chore: padding for some UI elements

* chore: physical mouse warning string

* refactor(theme): overhaul MaterialTheme color usage

* refactor(theme): shadow elevation & Card list color scheme

* chore: remove similar themes

* refactor(account): optimize serialization

* chore(TODO): Add TODO text to unfinished screens.

* update to 2.0.0_alpha-20250520a

* update gradle.properties

* fix: Quickly add snapshot version check when entering multiplayer mode

* chore(UI): tooltip's padding

---------

Co-authored-by: movte <movtery228@qq.com>